C++ Example
This application reads an external data table of time and temperature
values when heating 15, 20, and 25 circuit boards. The program then uses
linear regression to generate a linear equation. The program then
prompts the user to input a number of circuit boards and a desired
temperature, which will then calculate the time required to heat the
circuit boards.