HTML5/Javascript/CSS3 Example
This program generates a game that consists of a table of draggable
squares that are randomized in their location and a blank space. The
goal is to sort the squares from lowest value to highest value. When
this is achieved the screen will display a message stating that the user
has won the game.