Java Example
This program creates a payroll object that consists of a group of 7
employees. The user then inputs the employees hourly rate and hours
worked. The program then allows the user to display employee wages by
employee ID or their index in the array. The user can also modify the
data stored in the object.